finalise changes for PG 9.5 port; delay incompatible changes to caom-2.2 (s1859)

+++ caom2persistence/src/ca/nrc/cadc/caom2/harvester/state/HarvestSkipDAO.java
@@ -127,7 +127,7 @@ public class HarvestSkipDAO
             }
             sb.append(" ORDER BY lastModified ASC");
             
-            if (batchSize != null && batchSize.intValue() > 0)
+            if (batchSize != null && batchSize > 0)
                 sb.append(" LIMIT ").append(batchSize.toString());
             
             String sql = sb.toString();
